FPBX-25
The parking lot sync function in modify.php now pays attention to the clean_unknown_items flag.

+++ modify.php
@@ -1624,7 +1624,7 @@ function sync_conference_rooms() {
  *
  */
 function sync_parking_lot() {
-	global $coreServerId, $logger, $pest;
+	global $coreServerId, $logger, $pest, $serverInformation;
 
 	$logger->debug("Syncing parking lot");
 
@@ -1662,11 +1662,15 @@ function sync_parking_lot() {
 		}
 
 		/*
-		 * Remove all parking lots from the server that are not the default parking lot.
-		 * If the parking lot is not enabled remove all parking lots.
+		 * If the server has the default parking lot configured, and FreePBX has parking disabled,
+		 * remove the default parking lot. 
+		 * 
+		 * If other parking lots are found on the server, and clean unknown items
+		 * is enabled, remove those parking lots.
 		 */
 		foreach($serverParkingLotsAssoc as $parkingLotIdentifier => $parkingLot) {
-			if($parkingLotIdentifier != "default" || !$parkingEnabled) {
+			if(($parkingLotIdentifier == "default" && !$parkingEnabled) ||
+			   ($parkingLotIdentifier != "default" && $serverInformation['clean_unknown_items'] == '1')) {
 				$logger->debug("Removing parking lot: " . $parkingLotIdentifier);
 				$pest->delete("asterisk/" . $coreServerId . "/parkingLots/" . $parkingLot->id);
 				unset($serverParkingLotsAssoc[$parkingLotIdentifier]);
